+++ typescript/sdk/src/middleware/MiddlewareRouterDeployer.ts
@@ -1,0 +1,86 @@
+import { ethers } from 'ethers';
+
+import { ProxyAdmin__factory } from '@hyperlane-xyz/core';
+
+import { MultiProvider } from '../providers/MultiProvider';
+import { HyperlaneRouterDeployer } from '../router/HyperlaneRouterDeployer';
+import {
+  ProxiedRouterContracts,
+  RouterConfig,
+  RouterFactories,
+} from '../router/types';
+import { ChainMap, ChainName } from '../types';
+
+export abstract class MiddlewareRouterDeployer<
+  MiddlewareRouterConfig extends RouterConfig,
+  MiddlewareRouterContracts extends ProxiedRouterContracts,
+  MiddlewareFactories extends RouterFactories,
+> extends HyperlaneRouterDeployer<
+  MiddlewareRouterConfig,
+  MiddlewareRouterContracts,
+  MiddlewareFactories
+> {
+  constructor(
+    multiProvider: MultiProvider,
+    configMap: ChainMap<MiddlewareRouterConfig>,
+    factories: MiddlewareFactories,
+    protected create2salt = 'middlewarerouter',
+  ) {
+    super(multiProvider, configMap, factories);
+  }
+
+  constructorArgs(
+    _chain: ChainName,
+    _config: MiddlewareRouterConfig,
+  ): Parameters<MiddlewareFactories['router']['deploy']> {
+    return [] as any;
+  }
+
+  async initializeArgs(
+    chain: ChainName,
+    config: MiddlewareRouterConfig,
+  ): Promise<[string, string, string, string]> {
+    // configure owner as signer for additional initialization steps
+    // ownership is transferred to config.owner in HyperlaneRouterDeployer.deploy
+    const owner = await this.multiProvider.getSignerAddress(chain);
+    return [
+      config.mailbox,
+      config.interchainGasPaymaster,
+      config.interchainSecurityModule ?? ethers.constants.AddressZero,
+      owner,
+    ];
+  }
+
+  async deployContracts(
+    chain: ChainName,
+    config: MiddlewareRouterConfig,
+  ): Promise<MiddlewareRouterContracts> {
+    const proxyAdmin = await this.deployContractFromFactory(
+      chain,
+      new ProxyAdmin__factory(),
+      'proxyAdmin',
+      [],
+      { create2Salt: this.create2salt },
+    );
+    const initArgs = await this.initializeArgs(chain, config);
+    const proxiedRouter = await this.deployProxiedContract(
+      chain,
+      'router',
+      this.constructorArgs(chain, config),
+      proxyAdmin,
+      initArgs as any, // generic type inference fails here
+      {
+        create2Salt: this.create2salt,
+      },
+    );
+    await this.multiProvider.handleTx(
+      chain,
+      proxyAdmin.transferOwnership(config.owner),
+    );
+    return {
+      proxyAdmin,
+      proxiedRouter,
+      router: proxiedRouter.contract, // for backwards compatibility
+    } as any; // generic type inference fails here
+  }
+}
