+++ typescript/sdk/src/middleware/query/InterchainQueryDeployer.ts
@@ -1,0 +1,26 @@
+import { MultiProvider } from '../../providers/MultiProvider';
+import { RouterConfig } from '../../router/types';
+import { ChainMap } from '../../types';
+import { MiddlewareRouterDeployer } from '../MiddlewareRouterDeployer';
+
+import {
+  InterchainQueryContracts,
+  InterchainQueryFactories,
+  interchainQueryFactories,
+} from './contracts';
+
+type InterchainQueryConfig = RouterConfig;
+
+export class InterchainQueryDeployer extends MiddlewareRouterDeployer<
+  InterchainQueryConfig,
+  InterchainQueryContracts,
+  InterchainQueryFactories
+> {
+  constructor(
+    multiProvider: MultiProvider,
+    configMap: ChainMap<InterchainQueryConfig>,
+    create2salt = 'queryrouter',
+  ) {
+    super(multiProvider, configMap, interchainQueryFactories, create2salt);
+  }
+}
