--- typescript/sdk/src/router/app.ts
@@ -1,39 +1,0 @@
-import { BigNumber } from 'ethers';
-
-import { GasRouter } from '@hyperlane-xyz/core';
-import { types } from '@hyperlane-xyz/utils';
-
-import { HyperlaneApp } from '../HyperlaneApp';
-import { ChainMap, ChainName } from '../types';
-import { objMap, promiseObjAll } from '../utils/objects';
-
-import { RouterContracts } from './types';
-
-export class RouterApp<
-  Contracts extends RouterContracts,
-> extends HyperlaneApp<Contracts> {
-  getSecurityModules = (): Promise<ChainMap<types.Address>> =>
-    promiseObjAll(
-      objMap(this.contractsMap, (_, contracts) =>
-        contracts.router.interchainSecurityModule(),
-      ),
-    );
-
-  getOwners = (): Promise<ChainMap<types.Address>> =>
-    promiseObjAll(
-      objMap(this.contractsMap, (_, contracts) => contracts.router.owner()),
-    );
-}
-
-export class GasRouterApp<
-  Contracts extends RouterContracts<GasRouter>,
-> extends RouterApp<Contracts> {
-  async quoteGasPayment(
-    origin: ChainName,
-    destination: ChainName,
-  ): Promise<BigNumber> {
-    return this.getContracts(origin).router.quoteGasPayment(
-      this.multiProvider.getDomainId(destination),
-    );
-  }
-}
